style: 가독성 향상 위해 breadService에서 c,r,u,d 기능 분리되도록 코드 포맷팅
